// src/types/Ingredient.ts
export const enum IngredientType {
  Bun = "bun",
  Main = "main",
  Sauce = "sauce",
}

export type Ingredient = {
  _id: string;
  name: string;
  type: IngredientType;
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
  __v: number;
};